cleanup the api
